LayoutTests/platform/blackberry/http/tests/websocket/tests/hixie76/handshake-fail-by-cross-origin-expected.txt

Make sure Web Socket connection failed if origin mismatches.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

document.domain=localhost
ws.url=ws://localhost:8880/websocket/tests/fixed-origin
PASS connected is false
PASS origin is undefined.
PASS successfullyParsed is true

TEST COMPLETE
